+++ src/pydiverse/pipedag/backend/table/base.py
@@ -383,8 +383,8 @@ class BaseTableStore(TableHookResolver):
     def retrieve_all_task_metadata(self, task: MaterializingTask) -> list[TaskMetadata]:
         """Retrieves all metadata objects associated with a task from the store
 
-        As long as a metadata entry has the same task and stage name, it should
-        get returned.
+        As long as a metadata entry has the same task and stage name, as well
+        as the same position hash as the `task` object, it should get returned.
         """
 
     # Lazy Table Metadata