+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 
 - support execution of subflow
 - store final state of task in flow result object
+- tasks now have a `position_hash` associated with them to identify them purely based on their position (e.g. stage, name and input wiring) inside a flow.
 - breaking change to metadata: added position_hash to `tasks` metadata table and change type of hash columns from String(32) to String(20).
 
 0.4.1 (2023-06-17)